refactor: adjust import paths for types
