+++ src/modules/texts/PublicText/components/HeaderPublicText/HeaderPublicText.tsx
@@ -2,7 +2,7 @@ import { Button, Grid } from "@mui/material";
 import React from "react";
 import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
-import type { IAuthState } from "shared/common/store/Auth/types/Auth.types";
+import type { IAuthState } from "shared/common/store/Auth/types";
 import Logo from "shared/components/Logo";
 
 const HeaderPublicTextView: React.FC = () => {
