+++ src/shared/components/Drawer/components/DialogSearchAll/DialogSearchAll.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useDispatch, useSelector } from "react-redux";
 import { CLOSE_DIALOG_SEARCH_ALL_TYPE } from "shared/common/store/DialogSearchAll/DialogSearchAll.reduce";
-import { IDialogSearchAllState } from "shared/common/store/DialogSearchAll/types/DialogSearchAll.types";
+import { IDialogSearchAllState } from "shared/common/store/DialogSearchAll/types";
 import { createAction } from "shared/common/store/store.action";
 import DialogSearchAllView from "shared/components/Drawer/components/DialogSearchAll/DialogSearchAllView";
 import { DialogSearchAllSchema } from "shared/components/Drawer/components/DialogSearchAll/schema";
