+++ src/shared/components/Layout/Snackbar/Snackbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import useSnackbar from "shared/common/hooks/useSnackbar";
-import { ISnackbarState } from "shared/common/store/Snackbar/types/Snackbar.types";
+import { ISnackbarState } from "shared/common/store/Snackbar/types";
 import SnackbarView from "./SnackbarView";
 
 const THREE_SECONDS_AUTO_HIDE_DURATION = 3000;
