+++ src/shared/components/Session/Session.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { useLocation, useNavigate, useSearchParams } from "react-router-dom";
 import useLocalStorage from "shared/common/hooks/useLocalStorage";
 import { SIGNIN_TYPE } from "shared/common/store/Auth/Auth.reducer";
-import { IAuthState } from "shared/common/store/Auth/types/Auth.types";
+import { IAuthState } from "shared/common/store/Auth/types";
 import { createAction } from "shared/common/store/store.action";
 import api from "shared/services/api";
 import SessionView from "./SessionView";
